# Use AVX/AVX2 masks in `minmax_element` and `minmax` vectorization