@StephanTLavavej 2024-09-12
I've pushed a conflict-free merge with main to pick up the toolset update. No clang-format regen was necessary.